로그인 기반 AI 추천 공고 기능 개선 및 JWT 인증 연동 수정
backend:
- JobPostController.java
- JobPostService.java
- JobPostRecommendationService.java
- JobPostsDto.java
- TechStack.java

frontend:
- api.ts
- jobPostApi.ts
- AttentionSection.tsx

- 로그인 시 사용자 프로필 기반 AI 추천 공고 노출 기능 구현
- JWT 인증 구조 변경에 따라 JobPostController 인증 방식 수정
- JobPostsDto에 추천 점수 필드(recommendScore) 추가
- TechStack 직렬화 오류 방지를 위해 @JsonIgnore 추가
- JobPostRecommendationService 가중치 및 매칭 로직 개선
- JobPostService에서 추천 공고 서비스 연동 메서드 추가
- frontend api.ts에서 /recommended 엔드포인트 인증 처리 보정
- ai 추천 공고 전부 건너뛰기 한 사람은 모두가 추천하는 공고로 문구가 바뀜

+++ backend/src/main/java/com/we/hirehub/dto/support/JobPostsDto.java
@@ -45,7 +45,7 @@ public class JobPostsDto {
   private String hireType;
 
   // 🔥 AI 추천 점수 (추천 공고 정렬용)
-  private int recommendScore;
+  private Double recommendScore;
 
   /** Mini DTO */
   @Getter
